--- screens/README.md
@@ -1,9 +1,0 @@
-# Code Description
-Below is an outline of our code and how we structured the application
-
-- all navigation is in top level App.js
-- each individual .js file is its own component e.g. loading.js is loading screen, intro.js is intro screen, sleepDiary.js is sleep diary screen
-
-
-## Chatbot breakdown
-screens/chat-screen/home.js is where the chatbot screen lives
