+++ src/setting.js
@@ -1,0 +1,12 @@
+import React, { Component } from "react";
+import { StyleSheet, Text, View, Image } from "react-native";
+
+export default class Setting extends React.Component {
+  render() {
+    return (
+      <View style={{ flex: 1, justifyContent: "center", alignItems: "center" }}>
+        <Text>Settings!</Text>
+      </View>
+    );
+  }
+}
