+++ src/utils/helper-utils.js
@@ -1,0 +1,21 @@
+const getRandomAppState = appState => {
+  let items = Array.from(appState);
+  return items[Math.floor(Math.random() * items.length)];
+};
+
+const getRandomGenericTip = () => { //five options to randomly select from
+  let items = [1,2,3,4,5];
+  return items[Math.floor(Math.random() * items.length)];
+};
+
+const getNextAppState = appState => {
+  let items = Array.from(appState);
+  return items[0];
+};
+
+const getNextModule = ()=> { //three modules to randomly select from
+  let items = [1,2,3];
+  return items[Math.floor(Math.random() * items.length)];
+};
+
+export { getRandomAppState,getNextAppState,getRandomGenericTip,getNextModule };
